+++ CHANGELOG.md
@@ -51,6 +51,12 @@
 - add some icons
 # ReactControls (version: 0.2.182)
 # ReactControls (version: 0.2.187)
+# ReactControls (version: 0.2.216)
+
+## Bug Fixes
+
+- DataGrid: Notify when the row change.
+
 # ReactControls (version: 0.2.215)
 
 ## Refacto